added new grid layout -> switched from gridstack to react-grid-layout.
switched default MQTT channel to widgets.
