+++ package.json
@@ -34,6 +34,7 @@
     "react": "^16.14.0",
     "react-contextmenu": "^2.14.0",
     "react-dom": "^16.14.0",
+    "react-grid-layout": "^1.2.5",
     "react-modal": "^3.13.1",
     "react-redux": "^7.2.4",
     "react-split": "^2.0.11",
